// src/Base/utils/types.ts
export type DexType = 
  | "aerodromeslipstream" 
  | "uniswapv3" 
  | "pancakeswapv3"
  | "aerodrome"
  | "uniswapv2"
  | "sushiswapv3"
  | "baseswapv3"
  | "alienbasev3";